feat: Added CLI commands for cephfs replication
CMDs:
1. Enable CephFS Replication
2. Disable CephFS Replication
